test(server): relocate referenced-project run-prep tests to a dedicated file
Move the resolveRunReferencedProjects suite and the multi-project
workspace sync kill-switch test out of the issue service test file into
a dedicated heartbeat test file. The tests exercise heartbeat run-prep
exports, so they belong beside the other heartbeat tests, not with the
issue service tests. This change only moves tests; it adds no coverage
and changes no product code.
